stm: hom: fix warnings when compile for Orly
This patch fixes several warnings when compile for Orly

arch/arm/mach-stm/hom.c: In function ‘stm_hom_prepare_eram’:
arch/arm/mach-stm/hom.c:116:2: warning: passing argument 2 of ‘_memcpy_toio’ makes pointer from integer without a cast [enabled by default]
arch/arm/include/asm/io.h:199:13: note: expected ‘const void *’ but argument is of type ‘long int’
arch/arm/mach-stm/hom.c:127:2: warning: format ‘%x’ expects argument of type ‘unsigned int’, but argument 2 has type ‘void *’ [-Wformat]
arch/arm/mach-stm/hom.c:127:2: warning: format ‘%x’ expects argument of type ‘unsigned int’, but argument 3 has type ‘void *’ [-Wformat]
arch/arm/mach-stm/hom.c:127:2: warning: format ‘%x’ expects argument of type ‘unsigned int’, but argument 4 has type ‘void *’ [-Wformat]
arch/arm/mach-stm/hom.c: In function ‘stm_hom_enter’:
arch/arm/mach-stm/hom.c:161:26: warning: initialization makes integer from pointer without a cast [enabled by default]
arch/arm/mach-stm/hom.c:197:3: warning: passing argument 1 of ‘stm_hom_exec_on_eram’ makes pointer from integer without a cast [enabled by default]
arch/arm/mach-stm/hom.h:50:6: note: expected ‘struct stm_hom_eram_data *’ but argument is of type ‘long unsigned int’
arch/arm/mach-stm/hom.c: In function ‘stm_hom_register’:
arch/arm/mach-stm/hom.c:270:2: warning: passing argument 1 of ‘__arm_ioremap’ makes integer from pointer without a cast [enabled by default]
arch/arm/include/asm/io.h:83:22: note: expected ‘long unsigned int’ but argument is of type ‘void *’

+++ arch/arm/mach-stm/pm/hom-stih415.c
@@ -107,7 +107,7 @@ static int __init hom_stxh415_setup(void)
 
 	pr_info("[STM]: [PM]: [HoM]: Early console [%d] @ 0x%x\n",
 			stm_asc_console_device,
-			stxh415_hom.early_console_base);
+			(unsigned int) stxh415_hom.early_console_base);
 
 	ret = stm_hom_register(&stxh415_hom);
 	if (ret) {
